backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
cached-property==2.0.1
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
deepdiff==8.6.1
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.122.0
fastapi-cli==0.0.16
fastapi-cloud-cli==0.5.1
fastar==0.7.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
numpy==2.3.5
openai==2.8.1
orderly-set==5.5.0
pandas==2.3.3
portkey-ai==2.1.0
pybatfish==2025.7.7.2423
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.16.0
rignore==0.7.6
sentry-sdk==2.46.0
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typer==0.20.0
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
